refactor: rename validation file in utils file to match what it does exactly
