+++ src/controllers/timetable/deleteTimetable.ts
@@ -2,7 +2,7 @@ import { Request, Response } from "express";
 import { timetableRepository } from "../../repositories/timetableRepository";
 import { Timetable } from "../../entity/Timetable";
 import { z } from "zod";
-import { validate } from "../../utils/zodValidateBody";
+import { validate } from "../../utils/zodValidateRequest";
 import { User } from "../../entity/User";
 import { userRepository } from "../../repositories/userRepository";
